+++ nengo/ensemble.py
@@ -108,8 +108,8 @@ class Ensemble(NengoObject):
 
     probeable = ('decoded_output', 'input', 'scaled_encoders')
 
-    n_neurons = IntParam('n_neurons', default=None, low=1)
-    dimensions = IntParam('dimensions', default=None, low=1)
+    n_neurons = IntParam('n_neurons', low=1)
+    dimensions = IntParam('dimensions', low=1)
     radius = NumberParam('radius', default=1.0, low=1e-10)
     encoders = DistOrArrayParam('encoders',
                                 default=UniformHypersphere(surface=True),